Move TransactionIntegration to the runtime package
It shouldn't have been exposed as API as it's related to config.
